+++ R/calc_solar_time.R
@@ -11,7 +11,7 @@
 #' @import dplyr
 #' @examples 
 #' local.time <- as.POSIXct('2016-05-27 12:00:00', tz='America/New_York')
-#' solar.time <- calc_solar_time(local.time, longitude=3)
+#' solar.time <- calc_solar_time(local.time, longitude=-74)
 #' @export
 calc_solar_time <- function(local.time, longitude) {